common module that is built and can be reference by the client and server packages, also copied by the Dockerfile
