+++ common/tsconfig.json
@@ -1,0 +1,17 @@
+{
+  "extends": "../tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./build",
+    "lib": ["ES5", "ES6"],
+    "target": "ES6",
+    "module": "ES6",
+    "moduleResolution": "bundler",
+    "isolatedModules": true,
+    "noEmit": false,
+    "declaration": true,
+    "baseUrl": "./src"
+  },
+  "include": [
+    "src"
+  ]
+}
