+++ package.json
@@ -5,7 +5,8 @@
   "license": "MIT",
   "workspaces":[
     "client",
-    "server"
+    "server",
+    "common"
   ],
   "scripts": {
     "postinstall": "husky install",
