+++ server/src/utils/types.ts
@@ -1,10 +1,11 @@
+import { Player } from "common/build/types";
 import { Request } from "express";
 
 export interface CustomReq<T> extends Request {
   body: T;
 }
 export interface PlayerRecord {
-  player: "white" | "black"; // TODO: need a "common" module for types
+  player: Player;
   gameRecordId: string;
 }
 export interface GameRecord {
